clean up clean up everybody clean up
