+++ src/main/scala/org/broadinstitute/dsde/firecloud/model/DataUse.scala
@@ -1,7 +1,11 @@
 package org.broadinstitute.dsde.firecloud.model
 
-import org.broadinstitute.dsde.rawls.model.AttributeName
+import org.broadinstitute.dsde.rawls.model.{Attribute, AttributeName, AttributeFormat, PlainArrayAttributeListSerializer}
+import org.broadinstitute.dsde.rawls.model.WorkspaceJsonSupport.AttributeNameFormat
 import spray.http.Uri
+import spray.json._
+import spray.json.DefaultJsonProtocol._
+import spray.json.JsValue
 
 object DataUse {
 
@@ -60,6 +64,31 @@ object DataUse {
       POA = r.POA.getOrElse(false),
       NCU = r.NCU.getOrElse(false))
   }
+
+  case class StructuredDataRequest(generalResearchUse: Boolean,
+                                   healthMedicalBiomedicalUseOnly: Boolean,
+                                   diseaseUseOnly: Array[String],
+                                   commercialUseProhibited: Boolean,
+                                   forProfitUseProhibited: Boolean,
+                                   methodsResearchProhibited: Boolean,
+                                   aggregateLevelDataProhibited: Boolean,
+                                   controlsUseProhibited: Boolean,
+                                   genderUseOnly: String,
+                                   pediatricResearchOnly: Boolean,
+                                   irbRequired: Boolean,
+                                   prefix: Option[String])
+
+  case class StructuredDataResponse(consentCodes: Array[String],
+                                    dulvn: Int,
+                                    prefix: String,
+                                    structuredUseRestriction: Map[AttributeName, Attribute]) {
+    def formatWithPrefix(): Map[String, JsValue] = {
+      implicit val impAttributeFormat = new AttributeFormat with PlainArrayAttributeListSerializer
+      Map(prefix + "consentCodes" -> consentCodes.toJson,
+        prefix + "dulvn" -> dulvn.toJson,
+        prefix + "structuredUseRestriction" -> structuredUseRestriction.toJson)
+    }
+  }
 }
 
 object ConsentCodes extends Enumeration {
@@ -86,3 +115,4 @@ object ConsentCodes extends Enumeration {
   val allPreviousDurFieldNames = duRestrictionFieldNames ++ Seq(DS, "RS-POP", "futureUseDate")
   val diseaseLabelsAttributeName: AttributeName = AttributeName.withLibraryNS(DS)
 }
+
