+++ src/test/scala/org/broadinstitute/dsde/firecloud/integrationtest/DataUseRestrictionSearchSpec.scala
@@ -137,11 +137,6 @@ class DataUseRestrictionSearchSpec extends FreeSpec with SearchResultValidation 
         assertDataUseRestrictions(searchResponse, DataUseRestriction(DS = DataUseRestrictionTestFixtures.diseaseValuesInts), DataUseRestrictionTestFixtures.diseaseValuesLabels.map("DS:" + _))
       }
 
-//      "RS-POP:non-empty list dataset should have values" in {
-//        val searchResponse = searchFor("RS-POP")
-//        assertDataUseRestrictions(searchResponse, DataUseRestriction(`RS-POP` = Seq("TERM-1", "TERM-2")))
-//      }
-
       "IRB dataset should be indexed as true" in {
         val searchResponse = searchFor("IRB-unique")
         assertDataUseRestrictions(searchResponse, DataUseRestriction(IRB = true), Seq("IRB"))
@@ -162,7 +157,6 @@ class DataUseRestrictionSearchSpec extends FreeSpec with SearchResultValidation 
             `RS-PD` = true,
             `RS-G` = true,
             `RS-FM` = true,
-            //`RS-POP` = Seq("TERM-1", "TERM-2"),
             IRB = true
           ),
           Seq("NPU", "RS-G", "NCU", "HMB", "NMDS", "RS-FM", "NCTRL", "RS-PD", "NAGR", "GRU", "IRB") ++
