+++ src/test/scala/org/broadinstitute/dsde/firecloud/service/LibraryServiceSpec.scala
@@ -439,7 +439,6 @@ class LibraryServiceSpec extends BaseServiceSpec with FreeSpecLike with LibraryS
         "NCTRL" -> JsBoolean(false),
         "GRU" -> JsBoolean(false),
         "HMB" -> JsBoolean(false),
-        //"RS-POP" -> JsArray(),
         "DS" -> JsArray()
       )
